+++ internal/api/handlers_auth.go
@@ -11,7 +11,7 @@ const (
 	serverVersion                = "0.2.0"
 	apiVersion                   = "1.7.0"
 	minimumClientAPIVersion      = "1.0.0"
-	eventAPIVersion              = "1.9.0"
+	eventAPIVersion              = "1.9.1"
 	minimumClientEventAPIVersion = "1.3.0"
 )